Fix truncate YMD SoQL API calls for dates

+++ common-bq/src/main/scala/com/socrata/bq/soql/bqreps/FloatingTimestampRep.scala
@@ -14,7 +14,7 @@ class FloatingTimestampRep extends BigQueryRep[SoQLType, SoQLValue] {
 
   override def SoQL(row: Seq[String]): SoQLValue = {
     if (row.head== null) SoQLNull
-    else SoQLFloatingTimestamp(new LocalDateTime(row.head.toLong, DateTimeZone.UTC))
+    else SoQLFloatingTimestamp(new LocalDateTime(row.head.toDouble.toLong, DateTimeZone.UTC))
   }
 
   override def jvalue(value: SoQLValue): JValue = {
